Added -g flag to VirtualLab launch to open scpeific software packages

+++ Scripts/VLPackages/Salome/SalomeExec.sh
@@ -12,7 +12,7 @@ while getopts "c:f:p:a:r:" options; do
 	  pypath="${OPTARG}"
 	  ;;  
 	  a)
-	  args="${OPTARG}"
+	  args="args:${OPTARG}"
 	  ;;
 	  r)
 	  runflag="${OPTARG}"
@@ -25,11 +25,13 @@ export PYTHONPATH=/home/ibsim/VirtualLab:$pypath:$PYTHONPATH
 tmpfile=$(mktemp) # make temp file to write port number to
 
 # run salome
-$cmd -$runflag --ns-port-log $tmpfile  $filepath args:$args
+$cmd -$runflag --ns-port-log $tmpfile  $filepath $args
 
-# get port number from tmpfile and kill salome on that port
-port=$(cat $tmpfile)
-$cmd kill $port
+if [ $runflag = 't' ] ; then
+	# get port number from tmpfile and kill salome on that port (only needed if GUI hasnt been opened)
+	port=$(cat $tmpfile)
+	$cmd kill $port
+fi
